Started working on second lighting pass to propagate light across neighbors.

+++ Assets/VektorVoxels/Chunks/HeightData.cs
@@ -1,9 +1,9 @@
 ﻿namespace VektorVoxels.Chunks {
     public struct HeightData {
-        public byte Value;
+        public int Value;
         public bool Dirty;
 
-        public HeightData(byte value, bool dirty) {
+        public HeightData(int value, bool dirty) {
             Value = value;
             Dirty = dirty;
         }
